Simplify `someip` module.
Removes a lot of unnecessary code.

Serialization will be handled in a future commit.
